+++ packages/browser-wallet/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+-   Added confirmation message after successfully connecting to a network, at Network Connect page.
 -   Updated Manage token page, search result. For UI consistency between mobile and web platforms, token name in search results was replaced with tokenId.
 
 ## 2.6.0